# tracing,sql: introduce BenchmarkTracing