--- a/pkg/sql/conn_executor_exec.go
+++ b/pkg/sql/conn_executor_exec.go
@@ -636,6 +636,7 @@ func (ex *connExecutor) execStmtInOpenState(
 	if !alreadyRecording && stmtTraceThreshold > 0 {
 		ctx, stmtThresholdSpan = createRootOrChildSpan(ctx, "trace-stmt-threshold", ex.transitionCtx.tracer)
 		stmtThresholdSpan.SetVerbose(true)
+		// TODO(irfansharif): How come this span is never Finish()-ed? #58721.
 	}
 
 	if err := ex.dispatchToExecutionEngine(ctx, p, res); err != nil {