@tbg 2021-01-11
This benchmark is only doing single-node operations, which benefit from the local grpc optimization and which in turn elides the KV-side creation of spans. Try using

https://github.com/cockroachdb/cockroach/blob/e9214f68130b6942990a04da197a1e514595b1df/pkg/bench/bench_test.go#L378-L389

which also exercises the multi-node case via `ForEachDB`, and in particular makes sure all gateways are used equally.

I think what I did when I played with the allocations was to make a `BenchmarkSelectNone` based off of

https://github.com/tbg/cockroach/blob/e9214f68130b6942990a04da197a1e514595b1df/pkg/bench/bench_test.go#L42-L45

(which is too simplistic b/c it doesn't hit KV at all) that point selected from an empty table. We should do the same for the simplest write workload possible (since writing creates different spans) to cover out bases, but I don't think there's too much of a need to cover larger read/write sets.