@tbg 2021-01-12
It looks like we now have `BenchmarkTracing` in two places, and that the one in `kv_test.go` should go away.