@tbg 2021-01-12
FWIW, I don't see the large change in Scan1:

```
tobias@td:~/go/src/github.com/cockroachdb/cockroach$ benchstat legacy.txt background.txt 
name                                           old time/op    new time/op    delta
Tracing/Cockroach/tracing=X/Scan1-16              196µs ± 2%     196µs ± 1%    ~     (p=0.690 n=5+5)
Tracing/MultinodeCockroach/tracing=X/Scan1-16     483µs ± 1%     478µs ± 1%    ~     (p=0.056 n=5+5)

name                                           old alloc/op   new alloc/op   delta
Tracing/Cockroach/tracing=X/Scan1-16             23.3kB ± 1%    23.4kB ± 1%    ~     (p=0.548 n=5+5)
Tracing/MultinodeCockroach/tracing=X/Scan1-16    73.1kB ± 1%    74.8kB ± 2%  +2.34%  (p=0.032 n=5+5)

name                                           old allocs/op  new allocs/op  delta
Tracing/Cockroach/tracing=X/Scan1-16                247 ± 0%       248 ± 1%    ~     (p=0.286 n=5+5)
Tracing/MultinodeCockroach/tracing=X/Scan1-16       762 ± 0%       767 ± 1%    ~     (p=0.333 n=5+5)
```

I ran

```
go test -benchtime 25000x -count 5 -bench BenchmarkTracing/Cockroach/tracing=/Scan1 ./pkg/bench/ | grep BenchmarkTracing | tee out
```